Endre aktivitetsplan hjelpetekster og skrive om til TS
